+++ src/pgmonkey/connections/postgres/normal_connection.py
@@ -68,7 +68,10 @@ class PGNormalConnection(PostgresBaseConnection):
         return self
 
     def __exit__(self, exc_type, exc_val, exc_tb):
-        if exc_type:
-            self.rollback()
-        else:
-            self.commit()
+        try:
+            if exc_type:
+                self.rollback()
+            else:
+                self.commit()
+        finally:
+            self.disconnect()